renaming udongle file to peripheral. More accurate description since it will deal with dongles, flash, and time machines/after8
